Updated altimeter tolerances
The FCOM section that stipulated altimeter tolerances has been
removed. I assume the check was moved into the normal procedures in the
FCOM, and that this was then picked up in a transfer into EOMB. The
altimeter versus airfield elevation check does, however, appear have
gone missing in the process.